requirements.txt
altair==5.5.0
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.4
extra-streamlit-components==0.1.71
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.4
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
narwhals==1.14.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
psycopg2==2.9.10
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlparse==0.5.2
streamlit==1.40.2
streamlit-antd-components==0.3.2
streamlit-authenticator==0.4.1
streamlit-elements==0.1.0
streamlit-option-menu==0.4.0
streamlit-searchbox==0.1.19
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0